refactor(browser): consolidate CDP helpers in browserService.js
Resolves the TODO in messagePlaywrightSync.js asking to extract shared
CDP helpers to avoid duplication. Centralizes cdpRequest, listCdpPages,
findOrOpenPage, isAuthPage, and evaluateOnPage in browserService.js so
messagePlaywrightSync, messageActions, messageTokenExtractor, and
googleOAuthAutoConfig share one implementation.

- Adds optional timeout to evaluateOnPage and uses it to replace
  triggerPageRequest in messageTokenExtractor (was a near-verbatim copy).
- Existing browserService functions (navigateToUrl, getOpenPages,
  getCdpVersion) now build on cdpRequest instead of inlining fetch+host
  logic.
- messagePlaywrightSync re-exports the helpers as a compat shim so
  existing consumers and test mocks keep working unchanged.
- Adds named constants for the 10s default and 60s evaluate timeouts.

Net -34 lines across the three files. All 2749 server tests pass.
